_layouts/test.html

{% if page.svg %}<?xml version="1.0" encoding="UTF-8"?>
{% endif %}<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en"{% if page.svg %}
    xmlns:svg="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"{% endif %}>
<head>
    <title>PlotKit: {{ page.title }}</title>
    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />

    <link href="../stylesheets/tests.css" media="screen" rel="Stylesheet" type="text/css" />

    {% for stylesheet in page.test_stylesheets %}
    <link href="{{ stylesheet }}.css" media="screen" rel="Stylesheet" type="text/css" />
    {% endfor %}

    {% for script in page.custom_scripts %}
    <script src="{{ script }}" type="text/javascript"></script>
    {% endfor %}

    <script src="http://github.com/ept/plotkit/raw/master/PlotKit/MochiKit.js" type="text/javascript"></script>
    {% for script in page.scripts %}
    <script src="http://github.com/ept/plotkit/raw/master/PlotKit/{{ script }}.js" type="text/javascript"></script>
    {% endfor %}

    {% if page.unit_test %}
    <link href="../stylesheets/simple_test.css" media="screen" rel="Stylesheet" type="text/css" />
    <script src="http://svn.mochikit.com/mochikit/trunk/tests/SimpleTest/SimpleTest.js" type="text/javascript"></script>
    <script src="{{ page.unit_test }}.js" type="text/javascript"></script>
    {% endif %}

    {% for script in page.test_scripts %}
    <script src="{{ script }}.js" type="text/javascript"></script>
    {% endfor %}

    {% if page.svg %}
    <!-- Required for IE to support inlined SVG -->
    <object id="AdobeSVG" classid="clsid:78156a80-c6a1-4bbf-8e6a-3cd390eeb4e2" width="1" height="1"></object>
    <?import namespace="svg" implementation="#AdobeSVG"?>
    {% endif %}
</head>

<body>
    <div id="body">
    <h2>{{ page.title }}</h2>

    {{ content }}
    
    {% if page.unit_test %}
    <!-- unit testing borrowed from MochiKit -->
    <script type="text/javascript">
    try {
        //SimpleTest.saveScope(this);
        tests.{{ page.unit_test }}({ok:ok, is:is});
        //SimpleTest.verifyScope(this);
        //ok(true, "test suite finished!");
    } catch (err) {
        var s = "test suite failure!\n";
        var o = {};
        var k = null;
        for (k in err) {
            // ensure unique keys?!
            if (!o[k]) {
                s +=  k + ": " + err[k] + "\n";
                o[k] = err[k];
            }
        }
        ok(false, s);
    }
    </script>
    {% endif %}

    <div class="footer"><a href="http://www.liquidx.net/plotkit/">PlotKit</a></div>    
</body>
</html>
